task number 102

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -5,15 +5,15 @@
 */
 int main(void)
 {
-	int i = 1;
-	int j = 0;
-	long int fibo = 1;
+	long long int i = 1;
+	long long int j = 0;
+	long long int fibo = 1;
 	int counter = 0;
 
 	while (counter < 50)
 	{
 		fibo = i + j;
-		printf("%ld", fibo);
+		printf("%lld", fibo);
 		j = i;
 		i = fibo;
 		counter++;
